used database when no internet connection

+++ app/src/main/java/com/project/myproject/ui/viewmodels/UserViewModel.kt
@@ -12,6 +12,7 @@ import com.project.myproject.data.requests.LoginRequest
 import com.project.myproject.data.repository.Repository
 import com.project.myproject.data.repository.UserRepository
 import com.project.myproject.data.requests.AddContactRequest
+import com.project.myproject.utils.NetworkUtil
 import com.project.myproject.utils.SessionManager
 import com.project.myproject.utils.SettingPreference
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -31,7 +32,8 @@ class UserViewModel @Inject constructor(
     private val contactRepository: ContactRepository,
     private val userRepository: UserRepository,
     private val sessionManager: SessionManager,
-    private val settingPreference: SettingPreference
+    private val settingPreference: SettingPreference,
+    private val networkUtil: NetworkUtil
 ) : ViewModel() {
 
     private val _loginState = MutableStateFlow<LoginState>(LoginState.Idle)
@@ -67,42 +69,44 @@ class UserViewModel @Inject constructor(
     }
 
     fun registerUser(email: String, password: String, isUserRemembered: Boolean) {
+        if (networkUtil.isInternetAvailable()) {
+            viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
 
-        viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
+                val result = repository.createUser(CreateRequest(email, password))
 
-            val result = repository.createUser(CreateRequest(email, password))
+                withContext(Dispatchers.Main) {
+                    if (result != null) {
+                        val responseBodyData = result.data
 
-            withContext(Dispatchers.Main) {
-                if (result != null) {
-                    val responseBodyData = result.data
+                        handleUserResponse(responseBodyData.user, responseBodyData.accessToken, responseBodyData.refreshToken, isUserRemembered)
 
-                    handleUserResponse(responseBodyData.user, responseBodyData.accessToken, responseBodyData.refreshToken, isUserRemembered)
+                        _registrationState.value = RegistrationState.Success
+                    } else {
+                        _registrationState.value = RegistrationState.InvalidRegisterData
+                    }
 
-                    _registrationState.value = RegistrationState.Success
-                } else {
-                    _registrationState.value = RegistrationState.InvalidRegisterData
+                    _registrationState.value = RegistrationState.Idle
                 }
-
-                _registrationState.value = RegistrationState.Idle
             }
         }
     }
 
     fun loginUser(email: String, password: String, isUserRemembered: Boolean) {
+        if (networkUtil.isInternetAvailable()) {
+            viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
 
-        viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
+                val result = repository.loginUser(LoginRequest(email, password))
 
-            val result = repository.loginUser(LoginRequest(email, password))
+                withContext(Dispatchers.Main) {
+                    if (result != null) {
+                        val responseBodyData = result.data
 
-            withContext(Dispatchers.Main) {
-                if (result != null) {
-                    val responseBodyData = result.data
+                        handleUserResponse(responseBodyData.user, responseBodyData.accessToken, responseBodyData.refreshToken, isUserRemembered)
 
-                    handleUserResponse(responseBodyData.user, responseBodyData.accessToken, responseBodyData.refreshToken, isUserRemembered)
-
-                    _loginState.value = LoginState.Success
-                } else {
-                    _loginState.value = LoginState.InvalidLoginData
+                        _loginState.value = LoginState.Success
+                    } else {
+                        _loginState.value = LoginState.InvalidLoginData
+                    }
                 }
             }
         }
@@ -119,17 +123,18 @@ class UserViewModel @Inject constructor(
     }
 
     fun editUserNameAndPhone(userName: String, phoneNumber: String) {
+        if (networkUtil.isInternetAvailable()) {
+            viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
 
-        viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
+                val result = repository.editUser(sessionManager.getId(),
+                    EditUserRequest(userName, phoneNumber)
+                )
 
-            val result = repository.editUser(sessionManager.getId(),
-                EditUserRequest(userName, phoneNumber)
-            )
-
-            withContext(Dispatchers.Main) {
-                if (result != null) {
-                    currentUser = result.data.user
-                    _userEdited.value = true
+                withContext(Dispatchers.Main) {
+                    if (result != null) {
+                        currentUser = result.data.user
+                        _userEdited.value = true
+                    }
                 }
             }
         }
@@ -155,33 +160,48 @@ class UserViewModel @Inject constructor(
                     true
                 )
 
-                val result = repository.getUser(userId)
+                if (networkUtil.isInternetAvailable()) {
+                    val result = repository.getUser(userId)
 
-                withContext(Dispatchers.Main) {
-                    if (result != null) {
-                        currentUser = result.data.user
+                    withContext(Dispatchers.Main) {
+                        if (result != null) {
+                            currentUser = result.data.user
 
-                        _registrationState.value = RegistrationState.RememberedUser
-                        _loginState.value = LoginState.Success
+                            _registrationState.value = RegistrationState.RememberedUser
+                            _loginState.value = LoginState.Success
+                        }
                     }
+                } else {
+                    currentUser = userRepository.getUser(userId)
+
+                    _registrationState.value = RegistrationState.RememberedUser
+                    _loginState.value = LoginState.Success
                 }
             }
         }
     }
 
     fun fetchContacts() {
-
         viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
-            _loading.value = true
+            if (networkUtil.isInternetAvailable()) {
 
-            val result = repository.getUserContacts(sessionManager.getId())
+                _loading.value = true
 
-            withContext(Dispatchers.Main) {
-                _loading.value = false
-                if (result != null) {
-                    _contacts.value = UserToContactMapper.map(result.data.contacts)
-                    _contactsId.value = result.data.contacts.map { it.id }
-                    contactRepository.replaceContacts(_contacts.value)
+                val result = repository.getUserContacts(sessionManager.getId())
+
+                withContext(Dispatchers.Main) {
+                    _loading.value = false
+                    if (result != null) {
+                        _contacts.value = UserToContactMapper.map(result.data.contacts)
+                        _contactsId.value = result.data.contacts.map { it.id }
+                        contactRepository.replaceContacts(_contacts.value)
+                    }
+                }
+            } else {
+                _loading.value = true
+                withContext(Dispatchers.Main) {
+                    _contacts.value = contactRepository.getAllContacts()
+                    _loading.value = false
                 }
             }
         }
@@ -190,15 +210,24 @@ class UserViewModel @Inject constructor(
     fun fetchUsers() {
 
         viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
-            _loading.value = true
+            if (networkUtil.isInternetAvailable()) {
+                _loading.value = true
 
-            val result = repository.getAllUsers()
+                val result = repository.getAllUsers()
 
-            withContext(Dispatchers.Main) {
-                _loading.value = false
-                if (result != null) {
-                    _users.value = result.data.users.filter { it.id != sessionManager.getId() }
-                    userRepository.replaceUsers(_users.value)
+                withContext(Dispatchers.Main) {
+                    _loading.value = false
+                    if (result != null) {
+                        val fetchedUsers = result.data.users
+                        _users.value = fetchedUsers.filter { it.id != sessionManager.getId() }
+                        userRepository.replaceUsers(fetchedUsers)
+                    }
+                }
+            } else {
+                _loading.value = true
+                withContext(Dispatchers.Main) {
+                    _users.value = userRepository.getAllUsers().filter { it.id != sessionManager.getId() }
+                    _loading.value = false
                 }
             }
         }
@@ -207,13 +236,14 @@ class UserViewModel @Inject constructor(
     fun deleteContact(contactId: Int) {
 
         viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
+            if (networkUtil.isInternetAvailable()) {
+                val result = repository.deleteUserContact(sessionManager.getId(), contactId)
 
-            val result = repository.deleteUserContact(sessionManager.getId(), contactId)
-
-            withContext(Dispatchers.Main) {
-                if (result != null) {
-                    _contacts.value = UserToContactMapper.map(result.data.contacts)
-                    _contactsId.value = result.data.contacts.map { it.id }
+                withContext(Dispatchers.Main) {
+                    if (result != null) {
+                        _contacts.value = UserToContactMapper.map(result.data.contacts)
+                        _contactsId.value = result.data.contacts.map { it.id }
+                    }
                 }
             }
         }
@@ -222,16 +252,18 @@ class UserViewModel @Inject constructor(
     fun addContact(contactId: Int) {
 
         viewModelScope.launch(Dispatchers.IO + exceptionHandler) {
-            _loading.value = true
+            if (networkUtil.isInternetAvailable()) {
+                _loading.value = true
 
-            val result = repository.addContact(sessionManager.getId(), AddContactRequest(contactId))
+                val result = repository.addContact(sessionManager.getId(), AddContactRequest(contactId))
 
-            withContext(Dispatchers.Main) {
-                _loading.value = false
-                if (result != null) {
-                    _contacts.value = UserToContactMapper.map(result.data.contacts)
-                    _contactsId.value = result.data.contacts.map { it.id }
-                    _contactAdded.value = true
+                withContext(Dispatchers.Main) {
+                    _loading.value = false
+                    if (result != null) {
+                        _contacts.value = UserToContactMapper.map(result.data.contacts)
+                        _contactsId.value = result.data.contacts.map { it.id }
+                        _contactAdded.value = true
+                    }
                 }
             }
         }
